+++ src/lib.rs
@@ -284,6 +284,7 @@ use std::ops::{
     Add, AddAssign, BitAnd, BitAndAssign, BitOr, BitOrAssign, Deref, DerefMut, Index, IndexMut,
     Sub, SubAssign,
 };
+//mod unit_tests;
 
 type CounterMap<T, N> = HashMap<T, N>;
 
@@ -334,6 +335,7 @@ where
     N: Zero,
 {
     /// Create a new, empty `Counter`
+    #[must_use]
     pub fn new() -> Counter<T, N> {
         Counter {
             map: HashMap::new(),
@@ -1306,428 +1308,5 @@ where
             let entry = self.map.entry(item.clone()).or_insert_with(N::zero);
             *entry += item_count.clone();
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use maplit::hashmap;
-    use rand::Rng;
-    use std::collections::HashMap;
-
-    #[test]
-    fn test_creation() {
-        let _: Counter<usize> = Counter::new();
-
-        let initializer = &[1];
-        let counter = Counter::init(initializer);
-
-        let mut expected = HashMap::new();
-        static ONE: usize = 1;
-        expected.insert(&ONE, 1);
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_update() {
-        let mut counter = Counter::init("abbccc".chars());
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 2,
-            'c' => 3,
-        };
-        assert!(counter.map == expected);
-
-        counter.update("aeeeee".chars());
-        let expected = hashmap! {
-            'a' => 2,
-            'b' => 2,
-            'c' => 3,
-            'e' => 5,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_add_update_iterable() {
-        let mut counter = Counter::init("abbccc".chars());
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 2,
-            'c' => 3,
-        };
-        assert!(counter.map == expected);
-
-        counter += "aeeeee".chars();
-        let expected = hashmap! {
-            'a' => 2,
-            'b' => 2,
-            'c' => 3,
-            'e' => 5,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_add_update_counter() {
-        let mut counter = Counter::init("abbccc".chars());
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 2,
-            'c' => 3,
-        };
-        assert!(counter.map == expected);
-
-        let other = Counter::init("aeeeee".chars());
-        counter += other;
-        let expected = hashmap! {
-            'a' => 2,
-            'b' => 2,
-            'c' => 3,
-            'e' => 5,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_subtract() {
-        let mut counter = Counter::init("abbccc".chars());
-        counter.subtract("bbccddd".chars());
-        let expected = hashmap! {
-            'a' => 1,
-            'c' => 1,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_sub_update_iterable() {
-        let mut counter = Counter::init("abbccc".chars());
-        counter -= "bbccddd".chars();
-        let expected = hashmap! {
-            'a' => 1,
-            'c' => 1,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_sub_update_counter() {
-        let mut counter = Counter::init("abbccc".chars());
-        let other = Counter::init("bbccddd".chars());
-        counter -= other;
-        let expected = hashmap! {
-            'a' => 1,
-            'c' => 1,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_composite_add_sub() {
-        let mut counts = Counter::<_>::init(
-            "able babble table babble rabble table able fable scrabble".split_whitespace(),
-        );
-        // add or subtract an iterable of the same type
-        counts += "cain and abel fable table cable".split_whitespace();
-        // or add or subtract from another Counter of the same type
-        let other_counts = Counter::init("scrabble cabbie fable babble".split_whitespace());
-        let _diff = counts - other_counts;
-    }
-
-    #[test]
-    fn test_most_common() {
-        let counter = Counter::init("abbccc".chars());
-        let by_common = counter.most_common();
-        let expected = vec![('c', 3), ('b', 2), ('a', 1)];
-        assert!(by_common == expected);
-    }
-
-    #[test]
-    fn test_most_common_tiebreaker() {
-        let counter = Counter::init("eaddbbccc".chars());
-        let by_common = counter.most_common_tiebreaker(|&a, &b| a.cmp(&b));
-        let expected = vec![('c', 3), ('b', 2), ('d', 2), ('a', 1), ('e', 1)];
-        assert!(by_common == expected);
-    }
-
-    #[test]
-    fn test_most_common_tiebreaker_reversed() {
-        let counter = Counter::init("eaddbbccc".chars());
-        let by_common = counter.most_common_tiebreaker(|&a, &b| b.cmp(&a));
-        let expected = vec![('c', 3), ('d', 2), ('b', 2), ('e', 1), ('a', 1)];
-        assert!(by_common == expected);
-    }
-
-    // The main purpose of this test is to see that we can call `Counter::most_common_tiebreaker()`
-    // with a closure that is `FnMut` but not `Fn`.
-    #[test]
-    fn test_most_common_tiebreaker_fn_mut() {
-        let counter: Counter<_> = Counter::init("abracadabra".chars());
-        // Count how many times the tiebreaker closure is called.
-        let mut num_ties = 0;
-        let sorted = counter.most_common_tiebreaker(|a, b| {
-            num_ties += 1;
-            a.cmp(b)
-        });
-        let expected = vec![('a', 5), ('b', 2), ('r', 2), ('c', 1), ('d', 1)];
-        assert_eq!(sorted, expected);
-        // We should have called the tiebreaker twice: once to resolve the tie between `'b'` and
-        // `'r'` and once to resolve the tie between `'c'` and `'d'`.
-        assert_eq!(num_ties, 2);
-    }
-
-    #[test]
-    fn test_most_common_ordered() {
-        let counter = Counter::init("eaddbbccc".chars());
-        let by_common = counter.most_common_ordered();
-        let expected = vec![('c', 3), ('b', 2), ('d', 2), ('a', 1), ('e', 1)];
-        assert!(by_common == expected);
-    }
-
-    #[test]
-    fn test_k_most_common_ordered() {
-        let counter: Counter<_> = "abracadabra".chars().collect();
-        let all = counter.most_common_ordered();
-        for k in 0..=counter.len() {
-            let topk = counter.k_most_common_ordered(k);
-            assert_eq!(&topk, &all[..k]);
-        }
-    }
-
-    /// This test is fundamentally the same as `test_k_most_common_ordered`, but it operates on
-    /// a wider variety of data. In particular, it tests both longer, narrower, and wider
-    /// distributions of data than the other test does.
-    #[test]
-    fn test_k_most_common_ordered_heavy() {
-        let mut rng = rand::thread_rng();
-
-        for container_size in [5, 10, 25, 100, 256] {
-            for max_value_factor in [0.25, 0.5, 1.0, 1.25, 2.0, 10.0, 100.0] {
-                let max_value = ((container_size as f64) * max_value_factor) as u32;
-                let mut values = vec![0; container_size];
-                for value in values.iter_mut() {
-                    *value = rng.gen_range(0..=max_value);
-                }
-
-                let counter: Counter<_> = values.into_iter().collect();
-                let all = counter.most_common_ordered();
-                for k in 0..=counter.len() {
-                    let topk = counter.k_most_common_ordered(k);
-                    assert_eq!(&topk, &all[..k]);
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_total() {
-        let counter = Counter::init("".chars());
-        let total: usize = counter.total();
-        assert_eq!(total, 0);
-
-        let counter = Counter::init("eaddbbccc".chars());
-        let total: usize = counter.total();
-        assert_eq!(total, 9);
-    }
-
-    #[test]
-    fn test_add() {
-        let d = Counter::<_>::init("abbccc".chars());
-        let e = Counter::<_>::init("bccddd".chars());
-
-        let out = d + e;
-        let expected = Counter::init("abbbcccccddd".chars());
-        assert!(out == expected);
-    }
-
-    #[test]
-    fn test_sub() {
-        let d = Counter::<_>::init("abbccc".chars());
-        let e = Counter::<_>::init("bccddd".chars());
-
-        let out = d - e;
-        let expected = Counter::init("abc".chars());
-        assert!(out == expected);
-    }
-
-    #[test]
-    fn test_intersection() {
-        let d = Counter::<_>::init("abbccc".chars());
-        let e = Counter::<_>::init("bccddd".chars());
-
-        let out = d & e;
-        let expected = Counter::init("bcc".chars());
-        assert!(out == expected);
-    }
-
-    #[test]
-    fn test_union() {
-        let d = Counter::<_>::init("abbccc".chars());
-        let e = Counter::<_>::init("bccddd".chars());
-
-        let out = d | e;
-        let expected = Counter::init("abbcccddd".chars());
-        assert!(out == expected);
-    }
-
-    #[test]
-    fn test_delete_key_from_backing_map() {
-        let mut counter = Counter::<_>::init("aa-bb-cc".chars());
-        counter.remove(&'-');
-        assert!(counter == Counter::init("aabbcc".chars()));
-    }
-
-    #[test]
-    fn test_from_iter_simple() {
-        let counter = "abbccc".chars().collect::<Counter<_>>();
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 2,
-            'c' => 3,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_from_iter_tuple() {
-        let items = [('a', 1), ('b', 2), ('c', 3)];
-        let counter = items.iter().cloned().collect::<Counter<_>>();
-        let expected: HashMap<char, usize> = items.iter().cloned().collect();
-        assert_eq!(counter.map, expected);
-    }
-
-    #[test]
-    fn test_from_iter_tuple_with_duplicates() {
-        let items = [('a', 1), ('b', 2), ('c', 3)];
-        let counter = items
-            .iter()
-            .cycle()
-            .take(items.len() * 2)
-            .cloned()
-            .collect::<Counter<_>>();
-        let expected: HashMap<char, usize> = items.iter().map(|(c, n)| (*c, n * 2)).collect();
-        assert_eq!(counter.map, expected);
-    }
-
-    #[test]
-    fn test_extend_simple() {
-        let mut counter = "abbccc".chars().collect::<Counter<_>>();
-        counter.extend("bccddd".chars());
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 3,
-            'c' => 5,
-            'd' => 3,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_extend_tuple() {
-        let mut counter = "bccddd".chars().collect::<Counter<_>>();
-        let items = [('a', 1), ('b', 2), ('c', 3)];
-        counter.extend(items.iter().cloned());
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 3,
-            'c' => 5,
-            'd' => 3,
-        };
-        assert_eq!(counter.map, expected);
-    }
-
-    #[test]
-    fn test_extend_tuple_with_duplicates() {
-        let mut counter = "ccc".chars().collect::<Counter<_>>();
-        let items = [('a', 1), ('b', 2), ('c', 3)];
-        counter.extend(items.iter().cycle().take(items.len() * 2 - 1).cloned());
-        let expected: HashMap<char, usize> = items.iter().map(|(c, n)| (*c, n * 2)).collect();
-        assert_eq!(counter.map, expected);
-    }
-
-    #[test]
-    fn test_count_minimal_type() {
-        #[derive(Debug, Hash, PartialEq, Eq)]
-        struct Inty {
-            i: usize,
-        }
-
-        impl Inty {
-            pub fn new(i: usize) -> Inty {
-                Inty { i }
-            }
-        }
-
-        // <https://en.wikipedia.org/wiki/867-5309/Jenny>
-        let intys = vec![
-            Inty::new(8),
-            Inty::new(0),
-            Inty::new(0),
-            Inty::new(8),
-            Inty::new(6),
-            Inty::new(7),
-            Inty::new(5),
-            Inty::new(3),
-            Inty::new(0),
-            Inty::new(9),
-        ];
-
-        let inty_counts = Counter::init(intys);
-        // println!("{:?}", inty_counts.map); // test runner blanks this
-        // {Inty { i: 8 }: 2, Inty { i: 0 }: 3, Inty { i: 9 }: 1, Inty { i: 3 }: 1,
-        //  Inty { i: 7 }: 1, Inty { i: 6 }: 1, Inty { i: 5 }: 1}
-        assert!(inty_counts.map.get(&Inty { i: 8 }) == Some(&2));
-        assert!(inty_counts.map.get(&Inty { i: 0 }) == Some(&3));
-        assert!(inty_counts.map.get(&Inty { i: 6 }) == Some(&1));
-    }
-
-    #[test]
-    fn test_collect() {
-        let counter: Counter<_> = "abbccc".chars().collect();
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 2,
-            'c' => 3,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_non_usize_count() {
-        let counter: Counter<_, i8> = "abbccc".chars().collect();
-        let expected = hashmap! {
-            'a' => 1,
-            'b' => 2,
-            'c' => 3,
-        };
-        assert!(counter.map == expected);
-    }
-
-    #[test]
-    fn test_superset_non_usize_count() {
-        let mut a: Counter<_, i8> = "abbcccc".chars().collect();
-        let mut b: Counter<_, i8> = "abb".chars().collect();
-        assert!(a.is_superset(&b));
-        // Negative values are possible, a is no longer a superset
-        a[&'e'] = -1;
-        assert!(!a.is_superset(&b));
-        // Adjust b to make a a superset again
-        b[&'e'] = -2;
-        assert!(a.is_superset(&b));
-    }
-
-    #[test]
-    fn test_subset_non_usize_count() {
-        let mut a: Counter<_, i8> = "abb".chars().collect();
-        let mut b: Counter<_, i8> = "abbcccc".chars().collect();
-        assert!(a.is_subset(&b));
-        // Negative values are possible; a is no longer a subset
-        b[&'e'] = -1;
-        assert!(!a.is_subset(&b));
-        // Adjust a to make it a subset again
-        a[&'e'] = -2;
-        assert!(a.is_subset(&b));
     }
 }
